llm_project/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.9.0

# LangChain & LangGraph
langchain>=0.3.7
langchain-core>=0.3.17
langchain-community>=0.3.7
langchain-google-genai>=2.0.4
langchain-ollama>=0.2.0
langgraph>=0.2.38

# Vector DB & Embeddings
chromadb>=0.5.15
langchain-chroma>=0.1.4

# Document Loaders
pypdf>=5.0.0

# Web Search
tavily-python>=0.5.0
duckduckgo-search>=6.3.0

# Utilities
httpx>=0.27.0
aiofiles>=24.1.0
